Refactored node count approach ( nodes 175573 nps 719561 os darwin )

+++ src/uci.cpp
@@ -178,7 +178,7 @@ namespace sagittar {
                     ss << "cp " << (int) result.score;
                 }
                 ss << " depth " << (unsigned int) result.depth;
-                ss << " nodes " << (unsigned long long) result.nodes;
+                ss << " nodes " << (size_t) result.nodes;
                 ss << " time " << (unsigned long long) result.time;
                 ss << " hashfull " << (unsigned int) result.hashfull;
                 ss << " nps " << (unsigned long long) (result.nodes * 1000 / (result.time + 1));
